Code cleanup

+++ src/main/java/com/github/klikli_dev/occultism/client/render/tile/SacrificialBowlRenderer.java
@@ -28,7 +28,6 @@ import com.mojang.blaze3d.matrix.MatrixStack;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.IRenderTypeBuffer;
 import net.minecraft.client.renderer.ItemRenderer;
-import net.minecraft.util.math.vector.Quaternion;
 import net.minecraft.client.renderer.model.IBakedModel;
 import net.minecraft.client.renderer.model.ItemCameraTransforms;
 import net.minecraft.client.renderer.tileentity.TileEntityRenderer;
@@ -63,7 +62,7 @@ public class SacrificialBowlRenderer extends TileEntityRenderer<SacrificialBowlT
             //use system time to become independent of game time
             long systemTime = System.currentTimeMillis();
             //rotate item slowly around y axis
-            float angle = (systemTime / 16) % 360;
+            float angle = (float) (systemTime / 16) % 360;
             matrixStack.rotate(Vector3f.YP.rotationDegrees(angle));
 
             //Fixed scale
